refactor: removed SidebarItem since calling N times same hooks is pointless
